// src/api/types/breed.ts
type TBreedWeight = {
    imperial: string;
    metric: string;
}

type TImage = {
    id: string;
    url: string;
}

export type TBreed = {
    weight: TBreedWeight;
    id: string;
    name: string;
    cfa_url: string;
    vetstreet_url: string;
    vcahospitals_url: string;
    temperament: string;
    origin: string;
    country_codes: string;
    country_code: string;
    description: string;
    life_span: string;
    indoor: number;
    lap: number;
    adaptability: number;
    affection_level: number;
    child_friendly: number;
    cat_friendly: number;
    dog_friendly: number;
    energy_level: number;
    grooming: number;
    health_issues: number;
    intelligence: number;
    shedding_level: number;
    social_needs: number;
    stranger_friendly: number;
    vocalisation: number;
    bidability: number;
    experimental: number;
    hairless: number;
    natural: number;
    rare: number;
    rex: number;
    suppressed_tail: number;
    short_legs: number;
    wikipedia_url: string;
    hypoallergenic: number;
    reference_image_id: string;
    image: TImage;
}

export type TBreedById = {
    breeds: TBreed[];
    id: string;
    url: string;
    width: string;
    height: string;
}